feat(resharding): improved error handling (#10152)
- get rid of some unwraps in the resharding implementation
- implemented an integration test for the error handling


I'm yet to see what actually happens in a real node when such an error
occurs. This PR is just a good first step, making sure that the error is
correctly propagated to catchup. Ideally I'd like to report any errors
in the logs and in the metrics so that the node owner can try to
recover.