fix(pytests): fix rpc_finality.py (#10148)
https://github.com/near/nearcore/pull/9644 added a new RPC method for
sending transactions, and added unused "wait_until" arguments to the
existing ones. This change was actually not a no-op for the
"broadcast_tx_commit" method, since now that method will wait until all
transactions and descendant receipts are final, whereas the previous
behavior was closer to the new "INCLUDED" "wait_until" parameter, and we
used to just wait for the receipt outcomes to show up, but would return
before they're executed and finalized.

So fix it by making send_tx_and_wait() call the send_tx method to get
something closer to the old behavior.

Also, an unrelated problem with this test is that the "consensus" change
to the config were put in the wrong place, so that measure to decrease
the flakiness of this test wasn't actually in place