test: SyncSnapshotHosts is propagated between PeerManagers
When one PeerManager receives a SyncSnapshotHosts message,
it should propagate it to all of the connected PeerManagers.
The message should spread through the network, eventually
reaching PeerManagers that aren't directly connected to the sender.

I hit the file descriptor limit with this test, so I reused the logic
from accounts_data.rs to increase the limit.