remove the currently omnipresent red mark
Considering in four weeks of always seeing red on all PRs, it has
obviously not worked to push us to actually fix the nayduck tests.

This failing means that we no longer see the actual value added by other
tests, and especially the green checkmark that would otherwise be a good
indicator that a PR is ready to be merged: we now need to manually check
and scroll the statuses list in order to figure out whether a PR is
actually ready for merge.

Currently, this is a net negative in developer experience. I’m not
questioning the nayduck problem, but the solution seemed wrong to me,
and time has shown it does not actually works.

So, I’m petitioning to remove this check from PRs (but keep it on master
CI), it would at least make it more livable for all the developers who
have literally no idea how to even check which nayduck tests are
failing, and whose code is unrelated to these tests anyway.