feat: add database analyse-gas-usage command
Add a command which will allow to analyse gas usage on the blockchain.
The command can be executed by running:
```bash
./neard database analyse-gas-usage
```
The command will look at gas used in all of the blocks,
shards and accounts, and print out an analysis report.

This commit just adds the command, the implementation will
be added in the following commits.