Implement analysing gas usage
For each block we want to gather information about how much gas was
used in every shard and account.

Gas is used when transactions and receipts are executed, and every
such execution produces an ExecutionOutcome, which is saved in the
database.
We can read all ExecutionOutcomes originating from a given shard
from the database, and get all of the needed information from there.
Every ExecutionOutcome contains the AccountId of the executor
and the amount of gas that was burned during the execution.
It's everything that is needed for analysing gas usage.

Usage from all blocks is merged into a single instace of GasUsageStats
and the interesting pieces of information are displayed for the user.