Make it possible to analyse a range of block heights
Add a flag that can be used to specify a range of blocks
to be analysed.

Now running:
```bash
./neard database analyse-gas-usage --from-block-height 120 --to-block-height 130
```

Will analyse 11 blocks with heights in range of [120, 130]

The logic of converting command line arguments to block iterators
has been extracted to the block_iterators module. In the future
it'll be possible to reuse it for another command that wants to
analyse some subset of blocks.