Calculate the optimal split for each shard
For each shard, calculate the optimal account that the shard could
be split at, so that gas usage on both sides of the split is similar.
An exact split isn't always possible, because it's possible for one account
to consume 40% of all gas, but the function tries its best to make it fair.

Information about optimal splits is displayed for the user.