fix: logs and traces (#10231)
No functional changes.

Improving logging and traces to enable investigation of the network wide
issue of validators missing chunks.

Logging changes are directed to making it easier to see why a node might
have missed a chunk or a partial encoded chunk.

Current state of investigation:
Validators are missing chunks, because they are in a syncing state.
Unclear why the validator nodes are in a syncing state.

Tested by running a testnet node.
The logs have a lot of `process_partial_encoded_chunk` logs, but nothing
is overly spammy.