Remove GasUsageInShard::used_gas_total
The field GasUsageInShard::used_gas_total keeps the total amount of
gas used in a shard. The problem is that keeping it in sync with the
rest of the struct requires special care and is prone to bugs.
A mistake in this logic has already caused one bug in calculating
shard splits. It's safer to get rid of the field and just calculate
the total amount of used gas when needed. It doesn't take that long
to calculate this value from scratch.