Improve GasUsageInShard::calculate_split()
The previous implementation of calculate_split() worked okayish,
but it had some faults.

The first fault was that it split the shard into two halves where the left
one consistend of accounts smaller than the boundary account. This doesn't match
the logic that NEAR uses for boundary accounts. In NEAR the left half includes
the boundary account, i.e it's <= boundary_acc, not < boundary_acc.
The new function uses the same division as NEAR: (left <= split_account), (right > split_account).

The second fault was that the old function looked for a split where gas_left >= gas_right.
It isn't easy to prove that this is the optimal one - why is the left half supposed to be bigger?
Let's implement it in a way that is easier to understand. The new calculate_split()
looks for a split that minimizes the difference between the two halves. This makes sense,
we want the halves to be as similar as possible.

The commit also adds unit tests to make sure that calculate_split() works correctly.