feat: add rayon::try_map_result
Add a function which does the same thing as `concurrency::rayon::try_map`, but operates
on `Result` instead of `Option`.

It calls the provided function on every item from the iterator and collects
the results into a Vec. On the first error it stops the execution.
When an error occurs it returns the items that were successfully processed along with the error.

It's needed because the return type of `SnapshotHostInfo::verify()` will soon be changed
from `Option` to `Result` and `try_map` can't be used to collect Results.

The implementation is a bit different from `try_map`. I managed to implement it using
only rayon iterator combinators, without the use of atomics. I think it's a bit cleaner
this way.