feat: limit the number of shards ids in SnapshotHostInfo
`SnapshotHostInfo` contains a list of shard ids for which
a peer has a snapshot. Under normal circumstances this list
is pretty small - the maximum size would be the total number of shards.
But a malicious peer could craft a `SnapshotHostInfo` message with
very large number of shards - millions of shard ids. This could
cause problems on the receiver node, so there has to be limit
on the number of shard ids in a message.

Let's limit the number of shard ids in a single message to MAX_SHARDS_PER_HOST.
It's a constant that describes how many shards a single peer can track.
A peer doesn't keep more state snapshots than the number of tracked shards,
so this limit is reasonable.

In an ideal world we could check the current number of shards and reject
messages that contain more shard ids than the current number, but sadly
this can't be implemented. The problem is that the receiving node might
be behind the rest of the blockchain, and the latest information just
isn't available, so it can't check what the current number of shards is.
We could reject messages in such situations, but this would lead to loss
of information.
Limiting the number of shard ids to a constant number is an okay alternative.