feat: add sharding::MAX_SHARDS_PER_HOST
Add a constant which specifies the maximum amount of shards
that a single host can track.

It will be used to limit the size of incoming messages which
contain information about each tracked shard id.
Currently when a node receives such message, it has no way to verify its
correctness. This is because the node might not have up to date information
about the blockchain, so it doesn't know how many shard ids should there be.

Without this limit the node would have to handle messages with hunderds
of thousands of shards ids, otherwise it might miss important information.