fix(sync): Count the number of syncs
It solves the problem of nodes going into header sync or body sync for just a few seconds. It's not a problem for an rpc node, but for a validator that is enough to miss production of a chunk or a block.
Short syncs are not shown in prometheus due to the large poll interval.