fix: left half of shard split should be less than split_account
When splitting a shard into two halves the shard is divided
into two on the split_account (boundary_account).
The split should look like this:
left < boundary_account
right >= boundary_account

Previously I misunderstood the semantics and implemented the split
as if it was left <= boundary and right > boundary. Fix it.