feat: use the < operator instead of cmp method to improve readability
The following code is currently used for comparing `boundary_account` and `account_id`:
```rust
if boundary_account.cmp(account_id) == Greater {
    ...
}
```

IMO it's a bit confusing - it isn't immediately obvious whether this means `boundary_account` > `account_id`
or `account_id` > `boundary_account`. I misread this line and because of that I made a mistake in #10240.

Let's change it to something that is easier to read:
```rust
if account_id < boundary_account {
    ...
}
```