use MAX_SHARDS_PER_SNAPSHOT_HOST_INFO instead of MAX_SHARDS_PER_HOST
The constant `MAX_SHARDS_PER_SNAPSHOT_HOST_INFO` is a better representation
for what this constant does. `MAX_SHARDS_PER_HOST` suggests that it limits
the number of shards, but it doesn't really do it, it just limits the number
of shards in a single network message.

I decided to go with an approach analogous to the one for AccountData cache.
It uses a constant called MAX_ACCOUNT_DATA_SIZE_BYTES to limit the size of
network messages. It's strictly a network protocol thing, with no meaning
for the rest of the code. Let's go with this approach instead of the previous one.