Make PeerManager truncate SnapshotHostInfo.shards when it exceeds MAX_SHARDS_PER_SNAPSHOT_HOST_INFO
When PeerManager is asked to send out a SnapshotHostInfo message with more shards than the limit
allows it can't just send it out, because other peers will ban it for abusive behavior.
In such situation let's truncate the list of shards by choosing a random subset of elements from it.
Choosing a random subset slightly increases the chances that other peers will receive snapshot info
for every shard.