refactor: move runtime tests into a separate file (#10274)
This PR moves NightshadeRuntime tests module into a separate file
without changing anything else.

See https://github.com/near/nearcore/issues/10275 for more context.


In case having tests as a separate file is a good idea for the following
reasons:
- these are mostly integration-style tests with heavy setup (see
`TestEnv`) and quite a lot of code, so to me it seems like it deserves
to be in a separate file
- the code is already in `mod.rs` file, so we don't need to create a
separate directory structure.