feat: add a time limit for processing transactions in produce_chunk
To produce a chunk the node fetches transactions from the transaction
pool and adds them to the chunk. Adding a transaction to the chunk
can be computationally expensive - the transaction is converted to
receipt, which burns gas. Because of that the number of transactions
has to be limited. The node will fetch transactions until some kind of
limit is reached, and then decide that the chunk is ready and publish it.

Currently we have a limit of gas that can be used for preparing transactions,
but there is a big problem with it - it doesn't take into account discarding
invalid transactions from the pool. Processing an invalid transaction doesn't
burn gas, so the node could spend a lot of time discarding invalid transactinos
without ever reaching the limit. This would make chunk production very slow and
the others might decide to just skip the chunk instead of waiting for it.

Let's add another limit to `prepare_transactions`: a time limit. When producing
a chunk the node will be allowed to add new transactions to the chunk for up to X seconds.
This will ensure that a chunk is produced quickly, even when there are many invalid transactions.

The limit is an `Option` because there might be cases where we don't want to use a time limit.
For example there might be tests for `prepare_transactions` which test the gas_limit. In such
tests we wouldn't want the time limit to get in the way, it would only make the tests more unreliable.