[resharding] Add pytest for checking RPC calls after resharding (#10296)
This PR adds a basic pytest to check if the RPC calls to tx endpoint
work after a resharding event.

Note that this test currently works under the assumption that the node
is tracking all shards and future improvements to the tests where we may
need to redirect the request to a node tracking the specific shard we
are interested in, would come later.

Resolving https://github.com/near/nearcore/issues/5047