feat: add produce_chunk_add_transactions_time_limit config option
Add a config option that allows to specify a time limit on adding transactions during
chunk production. This option limits the amount of time that a node will spend adding
transactions to the produced chunk.

It's good to make this configurable - the default value seems reasonable,
but somone might want to tweak it in the future. There's no ideal value here,
choosing a limit that's too small would lead to chunks with very few transactions,
while choosing a value that's too big would lead to timeouts and skipped chunks.