Move default values for ClientConfig to the ClientConfig module
The default values for `ClientConfig` vere defined in the `nearcore` crate.
Let's move them to the same module where `ClientConfig` is defined, `client_config.rs`
in `near-chain-configs` crate. This makes it possible to use the default values in other
crates without importing the entire `nearcore` crate. Defaults should be declared in the same
module as the config, it's cleaner this way.