feat: report what limited the output of prepare_transactions()
When a chunk is produced, Client calls `prepare_transactions()`,
which fetches transactions from the pool and adds them to the chunk.

`prepare_transactions()` adds transactions until some limit is hit,
a limit on gas, time, number of transactions, chunk size, etc.
Currently `prepare_transacations()` doesn't report which limit was hit,
it just returns a list of processed transactions. Let's add a way to access
this information, it's useful to know what's the bottleneck of chunk production.

The return type of `prepare_transactions()` is changed to a struct which
contains the list of prepared transactions and the limit which was hit.
This allows the caller to find out what limited the produced chunk.
In the future the information about hitting a limit can be exposed in the metrics.