feat: add a metric which counts the number of produced chunks that didn't fit all transactions
Add a metric which counts the number produced chunks where some transactions
from the transaciton pool didn't fit into the chunk due to a limit.
The metric has a label `limited_by`, which lets us know which limit was hit when producing the chunk.
This will help to find out what's the bottleneck of chunk production in scenarios of high chain congestion.