fix: remove flakiness in the test try_map_result_stops_on_error
`try_map_result` is a function which takes a list of tasks and runs
them on multiple threads producing task outputs until the first error occurs.

The test `try_map_result_stops_on_error` asks to execute an infinite list of tasks,
but the 100th task throws an error.

The test checked that the function produces outputs for the first
10 tasks, but this isn't always true. The first 10 tasks could end
up on a thread that is preempted and the tasks aren't executed before
the error is hit on the 100th task. Because of this the test is flaky.

I assumed that the first 10 tasks would be executed first, but this was wrong,
we can't assume anything here. Let's remove all checks that could be broken
by race conditions, no matter how unlikely. This makes the test weaker,
but at least it won't be flaky.