style: remove cyclic dependency between primitives & vm-runner (#10343)
Everything related to parameters and configurations has been moved into
a low-dependency `near-parameters` crate. Thus now any tests that depend
on the parameters can do so easily without introducing dependencies on
crates that may depend on the crate being tested in the first place
(such as `near-vm-runner`.)

One thing of note is that some VM structures pertinent to configuration
were moved out of `near-vm-runner` into `near-parameters` as well, so
that `near-parameters` is largely a self-contained crate for all things
protocol config.

Fixes #10339