lychee: only check checked-in files (#10379)
In my local environment lychee was crawling through the `target/`
directory, which contained a link to a PDF at `dl.acm.org`. Running the
test suite one too many times has got me a ban(ana)d off their site.

Conversely having checked lychee's code, it only ever checks links in
markdown and html files. So for instance whatever links we might have in
our code or docstrings can easily remain broken and we’d be
none-the-wiser.

For what it is worth at least for the time being the number of links
checked remains the same as before.