refactor(resharding): renaming resharding code from "state split" to "resharding"  (#10393)
The current naming is centered around "state split". It's really
confusing to newcomers and even experienced developers as it's not
obvious what spltting state pertains to. The new naming is centered
around "resharding" which should be clear to anyone. Moreover in the
future we may want to add more resharding ops, beyond splitting shards,
so this refactoring makes it more future proof.

Care needs to be taken when handling the configs. Ideally this PR should
be included in the same rollout as the resharding itself. Otherwise a
node operator may configure resharding using the legacy
"state_split_config" which will become obsolete as of this PR. cc
@posvyatokum