feat(amend-genesis): allow setting the gas price (#10389)
After https://github.com/near/NEPs/issues/92, the min gas price was set
to 100M yoctoNEAR. But the genesis file we get from the state viewer
dump-state command copies the 1B min gas price from the genesis file.
This results in a network with 10x the gas fees as mainnet when we
launch mocknet testing networks, and many transactions sent by the
mirror run command fail with too little balance. So this PR allows
setting the min/max gas prices in the mocknet genesis file so we can fix
that. It might also be worth changing the behavior of `view-state
dump-state` to set the min gas price, but this will work for now and
lets us get the right gas price in all cases