use bigger runners for CI
With this change, all checks should complete within 10m, with the
exception of the M1 MacOS one.

This should be enough to give a reasonably quick answer to PR
submitters, so that (especially newcomers who don’t have the whole
setup) they could quickly know whether the PR is ready to be accepted or
not.