store: increase the size of the CachedContractCode column cache
This possibly mitigates the RocksDB slowness of loading contracts from the storage
after the another in-memory cache fronting RocksDB was removed in #9244. As part of
the 1.35 release[^1] it turned out that this has caused a significant bimodality in
chunk application times – mostly attributed to storage layer. This, we suspected,
had started causing increased validator kick out numbers.

Original LRU cache fronting RocksDB stored 128 units of loaded contracts. At a rough
estimate of 4MiB of machine code per contract, this comes out to 512MiB. Of course
the two caches are still very different – this stores uncompressed blocks containing
compiled machine code. The previous cache stored contracts loaded into memory. I'm
confident that the overhead of loading the contract code into memory is not
particularly notable at this point, though, so the difference is something we can
live with.

[^1]: https://near.zulipchat.com/#narrow/stream/297873-pagoda.2Fnode/topic/release.201.2E36/near/379169243