[stateless_validation] Introduce versioned BlockBody (#10411)
This PR introduces versioned BlockBody struct which would be a part of
BlockV4. Note that this isn't being used with the current protocol
version.

Eventually with stateless validation, we are going to have chunk
endorsements as part of the BlockBody, the structure of which would look
something like this or equivalent (still undecided if we need
`Vec<Vec<ChunkEndorsement>>` or `Vec<ChunkEndorsement>`)

```
pub struct BlockBodyV2 {
    pub chunks: Vec<ShardChunkHeader>,
    pub challenges: Challenges,

    // Data to confirm the correctness of randomness beacon output
    pub vrf_value: Value,
    pub vrf_proof: Proof,
    
    // Chunk Endorsements
    pub chunk_endorsements: Vec<ChunkEndorsement>,
}
```