style: sever some dependencies between primitives and vm-runner
Most notably things like conversions between error types and re-exports
of e.g. TrieNodesCount.

There is still a dependency on a few instances however, but those are
much less clear cut on how to resolve. In particular at least
`ProfileDataV3` is in an icky place. It should be split up into a
serialized schema (just like the errors were) and the simple structure
that `near-vm-runner` produces with a conversion in some integration
crate downstream…