Add the same wallet_contract.wasm as on master
In a previous commit I remove wallet_contract.wasm
thinking that it shouldn't be tracked by git, but
it turns out that it should. Let's take wallet_contract.wasm
from master and add it to the git. The effect will be that
there will be no change to wallet_contract.wasm