[stateless_validation] Process and store incoming chunk endorsements to block produce (#10416)
This PR adds the implementation for `process_chunk_endorsement` function
in client.

When a client gets a new chunk endorsement, we do the following
- Verify if the chunk endorsement is from a valid chunk validator
- Verify if the signature is correct
- Store the chunk endorsement in-memory in a LRU cache
`LruCache<ChunkHash, HashMap<AccountId, ChunkEndorsement>>`

The ChunkEndorsements would later be included in the block body during
block production.

Minor cleanup to the ChunkEndorsement struct as well.