style: check for unused dependencies using cargo-udeps (#10433)
As evidenced by some of the changes, we have been pulling in a couple of
dependencies that aren’t being used in code in any way. While that does
not affect the runtime (dead crates don’t even get looked at by rustc,)
it very much so does affect the compile times.

One unfortunate part of this check is that running it forces a `cargo
check` of the entire worktree. That’s not ideal (and would be great if
there was an easy way to combine it with some other work,) but it might
do for now?