parameters: move out test strings to a separate file (#10439)
One of these strings intentionally has some trailing space at the end of
the line. This is pretty hard to work with when the editor is set up to
strip these off, as in my case, making modifying those files a chore.

Moving them into files that’ll never get touched as a common case solves
this problem altogether.