feat(state-sync): implement peer selection for state part requests
This implements the ideas laid out here: https://github.com/near/nearcore/issues/10213,
and implements a select_host() function for the SnapshotHostsCache. This function
selects a peer to request the part from based on the up-to-date SnapshotHostInfos
we've seen, and sorts them by a priority score calculated from the Peer Id and Part Id,
remembering who we've requested from so we can try different peers on subsequent calls

A goal of the implementation is to have as little impact as possible on nodes not
running state sync, which will be most nodes. So here we lazily only store data associated
with peer selection for a given part after select_host() has been called for that part, and
the added part_received() function is meant to serve as a signal that we can free this data