feat: add a metric which counts how many chunks couldn't fit all transactions from the pool (#10422)
When a chunk is produced, Client calls `prepare_transactions()`, which
fetches transactions from the transaction pool and adds them to the
chunk. `prepare_transactions()` adds transactions until some limit is
hit, a limit on gas, time, number of transactions, chunk size, etc.

Currently there's no way to know if some limit was hit and which limit
was it. Let's add a way to access this information, it's useful to know
what's the bottleneck of chunk production.

This PR adds a new metric which counts how many of the produced chunks
couldn't fit all transactions from the transaction pool due to hitting a
limit. It has two labels: `shard_id` - chunk's shard and `limited_by` -
which of the limits was hit in `prepare_transactions()`.

The need for this metric was discussed in
https://github.com/near/nearcore/pull/10310#issuecomment-1850163958. The
hope is that it'll allow us to figure out what's the bottleneck of chunk
production in scenarios of high chain congestion.
Right now in cases where `produce_chunk` takes a lot of time we can only
make theories why that's the case, but there's no observability into
what's going inside. This metric could help with that.

To test that the metric works I set up a mocknet network with a build
from this branch and put a load on it using locust.
I reduced `produce_chunk_add_transactions_time_limit` from 200ms to 2ms
to easily trigger the metric. It worked, as can be observed on the
grafana dashboard:
https://nearinc.grafana.net/d/eada7f01-b2dc-4df8-8a9f-ec4ec411159e/jancio-mocknet-stats?orgId=1&from=1705135200000&to=1705135800000
![Screenshot 2024-01-13 at 15 21
12](https://github.com/near/nearcore/assets/149345204/51ba6a04-c2ff-4773-a8d7-9cfbff8a7fa5)