fix codecov intermittent failure on PRs from forked repos
Forked repos do not actually have access to the codecov token, and
codecov is subject to rate-limiting which leads to intermittent
failures.

Making the token public is not a big deal, because we really treat all
information on codecov as informational only, and do not publicize it as
an important part of our system.