Pass prev_block to send_chunk_state_witness
Let's pass a prev_block to send_chunk_state_witness,
as it's needed to generate source_receipt_proofs.

We can remove the prev_chunk_header argument to keep
the number of arguments small. It requires us to recalculate
it, but it's fast and easy to do, so there's no harm in it.