fix(configs): fill in missing values in ExperimentalConfig with defaults
the command `neard --home {home_dir} init --chain-id mainnet --download-config`
downloads the config at
https://s3-us-west-1.amazonaws.com/build.nearprotocol.com/nearcore-deploy/mainnet/config.json
and then saves a config file built from that one to
{home_dir}/config.json. This is expected to work even when particular
fields are missing by filling them in with defaults, which is why many
config fields are marked with some sort of #[serde(default)] or
doesn't fill in these defaults. The existing #[serde(default)] over
that field in `near_network::config_json::Config` will have us fill it
in if it's totally missing, but we get an error if it's present with
some fields missing, which is the case today:

Error: Failed to initialize configs

Caused by:
    config.json file issue: Failed to deserialize config from /tmp/n/config.json: Error("missing field `network_config_overrides`", line: 98, column: 5)

Fix it by adding a #[serde(default)] to the ExperimentalConfig struct
itself, which is how, for example, the fact that the `archive` field
is missing in the above config url is not a problem, because
`nearcore::config::Config` is marked with this attribute.