fix: use prev_chunk_header instead of prev_chunk in collect_source_rceipt_proofs
It turns out that prev_chunk.height_included is sometimes 0, even though
the chunk was included at a nonzero height. This caused an error in collect_source_receipt_proofs,
because is_new_chunk was never true and the loop went on forever.
Let's use prev_chunk_header instead, height_included is always correct in the header.