[stateless_validaiton] Skip validation of state witness for chunk right after genesis (#10487)
We are skipping state witness validation for the chunk right after
genesis chunk due to some complications of trying to run the genesis
chunk in runtime (it's not possible to do so).

Initially this edge case check was a part of
`send_chunk_state_witness_to_chunk_validators` and we didn't send state
witness at all.

This has an issue that we do not process and send chunk endorsements
either due to this. Instead pushed this check into
`process_chunk_state_witness` and we send chunk endorsements for this
special case.

---------

Co-authored-by: Aleksandr Logunov <alex.logunov@near.org>