feat: sign chunk state witness (#10482)
It looks reasonable to have some protection against sending arbitrary
data pretending to be state witness, as it is large message causing a
lot of computation. Also having a message to reject of state witness is
more convenient for testing.

Here I modify state witness message to include signature; construct and
verify it; and also spawn message to ban peer if state witness is
invalid. As chunk producer is uniquely determined by state witness data,
we don't need to pass it in message. Ideally the test also should be
here but it is not a trivial change so I'll send it in the next PRs.

The logic of `SignatureDifferentiator` can be done better but for now I
just follow chunk endorsements pattern. Also I recognize that
signing/verifying logic is quite duplicated around epoch manager, but I
don't see an easy way around it.

## Testing

For now, just check that signing logic works.

---------

Co-authored-by: Longarithm <the.aleksandr.logunov@gmail.com>