feat: validate source_receipt_proofs in ChunkStateWitness
`source_receipt_proofs` describe the receipts that should be applied
to execute the state transition proven by `ChunkStateWitness`.

To validate them, we have to go over the chunks which have outgoing
receipts that should be applied during the state transition,
grab the right proof from `source_receipt_proofs` and validate that
the receipts in the proof really come from this chunk.

The receipts also have to be arranged in the order in which they
are going to be applied. I based the code mainly on the behavior in
`collect_incoming_receipts_from_block()` and `get_incoming_receipts_for_shard()`
to make sure that the receipts are in the correct order.