feat(state-sync): Download state header from external storage
In a previous PR, we have introduced the option to upload the state parts to external storage.
In this PR we are changing the way that state sync headers are requested, introducing to download
from external storage. The motivation behind this is that sometimes peers may not track all shards
making it impossible for them to share the state sync header.