refactor: add prev_ prefix to receipts in PartialEncodedChunk
The `receipts` field in `PartialEncodedChunk` represents the outgoing
receipts of the previous chunk. Let's modify the name to make it more
descriptive: `prev_outgoing_receipts` is much clearer than just `receipts`.
The same name is already used in `ShardChunk`, so it'll make the code more
consistent and readable.

The change shouldn't break anything, the reasoning is the same as in
https://github.com/near/nearcore/pull/9500:

> This change is safe since field names are not used for serialisation because those structs only support Borsh serialization.