Add EntityDebug UI entities for stateless validation (#10524)
Two new queries added:
- StateTransitionData: given block hash, query the StateTransitionData
stored for each shard at that block (we don't query by chunk hash here
because this data also exists for missing chunks).
- ValidatorAssignmentsAtHeight: given block height and epoch ID, query
the block producer, chunk producer for each shard, and the chunk
validator assignments.

Also add the ChunkEndorsements field to the Block result. Here, adding
it to the BlockView is probably better, but that is a larger change.