Add ProcessingDoneTracker and ProcessingDoneWaiter
Add two structs which can be used to wait until some kind of processing is done.
The ProcessingDoneTracker is paired with a ProcessingDoneWaiter,
and then the tracker can be passed to some function that runs in the background.
Calling ProcessingDoneWaiter::wait() will wait until the paired tracker is destroyed,
which happens when the function finishes processing.