fix(resharding): delay deleting the snapshot until after catchup (#10545)
I haven't verified it but it seem that restarting the node during
catchup does not resume the catchup but rather it restarts the whole
resharding from scratch. While suboptimal I'm not aiming to fix that
now. This PR merely moves the deletion of state snapshot to after
catchup is finished. This way the restarted resharding can succeed.