Add processing_done_tracker argument to process_chunk_state_witness
Add an additional argument to process_chunk_state_witness which allows
to wait until the processing is finished. process_chunk_state_witness
spawns a background task which processes the witness, because of which
it was hard to figure out when the witness has been processed.

With this argument, one can easily wait for the processing to finish.
The argument is optional, if the caller doesn't need to wait for it
to finish, they can just pass None.