Use ProcessingDoneTracker to propagate state witnesses and chunk endorsements in integration tests
Exchanging messages in integration tests has to be done manually,
the network stack isn't available there. It's quite hard to do it
correctly for chunk endoresements, because chunk endorsements are
sent out after a state witness has been processed, and state witnesses
are processed asynchronously in the background, and there was no way
to wait for the processing to finish.

The current approach is to define the set of chunk endorsements that should
be sent out and wait for the right messages to appear. This approach has some
limitations, it's often hard to figure out what this set should be, as the test
scenarios can be pretty convoluted, with missed/invalid chunks, skipped blocks, forks, etc.

It'd be much easier if we could wait until the state witnesses are processed,
and then process all chunk endorsements messages, without defining the expected set to wait for.

In the previous commits I added the `processing_done_tracker` argument to `process_chunk_state_witness`,
which makes this possible. Thanks to it, we can wait until the witnesses are processed, and then simply
handle all endorsement messages. This makes writing integration tests much simpler.