fix: test_banning_chunk_producer_when_seeing_invalid_[chunk|tx_in_chunk]
Fix two tests which test how the blockchain behaves when there is one malicious
chunk producer who produces invalid chunks. The tests didn't work with stateless validation.

To fix them we first have to properly propagate chunk state witnesses and chunk endorsements.
We should allow errors when processing witnesses and endorsements, as the malicious chunk producer
will cause some of the validations to fail.

Then it's also necessary to adjust the expected behavior. The previous version of the protocol
first included new chunks into a block and then validated the whole block. This means that if some
chunk was invalid, it would cause the whole block to be invalid, and the block would be skipped.
This is why the current code expected the first block with invalid chunks in an epoch to be skipped.
With stateless validation, the situation is different. Chunk validators will refuse to send endorsements
for the invalid chunk, which means that it won't be included in any blocks, and no blocks will be skipped.
The only exception is the block at height 2, which is skipped. I suspect that this is because the first few
blocks in the blockchain still use the old validation logic, not stateless validation.