[resharding] Fix opening snapshot for split storage nodes (#10588)
In testnet we hit an issue during resharding with split storage nodes
where they failed to create the snapshot required for resharding. This
PR is a fix for this issue.

Link to zulip thread:
https://near.zulipchat.com/#narrow/stream/308695-pagoda.2Fprivate/topic/cutting.201.2E37.2E0/near/420092380

Creation of snapshot happens via the function call
`checkpoint_hot_storage_and_cleanup_columns` which takes a hot_store and
creates a snapshot out of it. Later we open the snapshot
`opener.open_in_mode(Mode::ReadWriteExisting)`.

The function `open_in_mode` -> `ensure_kind` -> `is_valid_kind_archive`
was the point of failure. This is evident from the log line
```
Feb 06 14:19:29 testnet-rpc-archive-public-02-asia-east1-b-fc341f59 neard[1753]: 2024-02-06T14:19:29.215599Z ERROR state_snapshot: State snapshot creation failed err=Hot database kind should be RPC but got Some(Hot). Did you forget to set archive on your store opener?
```

There are three types of nodes, RPC, legacy archival and split storage
nodes.

The function call to `get_default_kind` gets us the DbKind for the new
snapshot hot storage while opening

To get the snapshot storage to open properly, we need to handle the
three types of storages. We need to set the value of `archive`
appropriately which is passed to the storage opener and ensure we set
the correct storage type of the snapshot.

The table below talks about the expected values of each of these fields.

| Node type | Hot storage kind | Required snapshot storage type |
archive |
| -- | -- | -- | -- |
| RPC | DbKind::RPC | RPC | false |
| Legacy Archival | DbKind::Archival | Legacy Archival | true |
| Split Storage | DbKind::Hot | RPC | false |

The place where our code went wrong with the split storage node was in
function call to `is_valid_kind_archive` where the DbKind is Hot for
split storage and archive is set as false.

The new relaxation in check basically states, if we are creating a
snapshot of the hot storage from a split storage node, convert it into a
RPC storage.

Testing: Adhoc testing where I manually set the storage DbKind as Hot
and check the fix works.
Additionally added an integration test that manually sets the DbKind as
Hot, Archive and RPC and checks whether resharding happens.