fix: make test_validate_chunk_extra work with stateless validation
This test didn't work with stateless validation. It manually creates a chunk
and blocks containing this chunk, but it didn't set the `chunk_endorsements`
field in the blocks, because of which the blocks couldn't be accepted.
Let's fix it by generating the correct chunk endoresment and setting the
field in the generated blocks.