fix(mirror): fix skipped account creations in the fifth block
When `mirror run` is run for the first time, we call
add_create_account_txs() for each of the first 5 blocks before we
start sending transactions because normally this is called on the
block that's 5 blocks ahead of the block we're currently preparing
transactions for. But there is an off-by-one error here, where we take
the first 5 block heights, say [10, 11, 12, 13, 14]. And then on the
first call to queue_txs(), `tracker.next_heights()` will return 10 for
the next height, and 16 for the height to look for create account
receipts, skipping 15. So we need to just add one to the number of
blocks we send create account txs for at the start of sending traffic