Improve the memtrie integration test to include a state sync base case. (#10592)
It turns out that even setting up a e2e test case for single-shard
tracking is tricky, so this PR does that. It makes a few fixes to the
test framework as well as the memtrie test code, so that
- In the track-all-shards + single-shard-tracking case (i.e. single
shard tracking is enabled but nodes still track all shards), allow test
code a chance to request partial encoded chunks.
- Query account balance from a client that tracks the shard (otherwise
we would get trie node missing errors).
- Send txns to every client instead of client 0, because not everyone
produces chunks for every shard now.
- Check that every block contains all new chunks, to self-verify that
the partial encoded chunk propagation is correct. Also check that the
total number of transactions included is correct in the end.