estimator: further improve error handling
Turns out the EBADF comes out regardless of whether the code runs inside
or outside of QEMU. However we can still get an indication of whether
the operations are successful or not by observing the buffer that we
pass in. Under QEMU it is now always overwritten with value usually
distinct from `u64::MAX`.