estimator: check binaries back in temporarily
The docker-based instructions rely on these binaries for the time being
and changing the docker image for now is a little bit over the top. I
would love for docker to reproducibly build these artifacts for itself,
but until that happens we will have to live with these binaries for a
little while longer.