Add possible_epochs_of_height_around_tip
Add a function which tries to determine in which epoch
the specified height is located.
It looks at the previous, current and next epoch around
the chain Tip and tries to figure out to which one of
them this height belongs.
It's not always possible to determine in which epoch
a given height will be, as the start height of the next
epoch isn't known in the current epoch, so the function
returns a set of possible epochs where the height might be.

For example if the tip is in the middle of the current epoch,
and the height is about epoch_length blocks higher than the tip,
this function would return [current_epoch_id, next_epoch_id],
as this height could be in either one of those epochs.

It will later be used to verify the signature of orphaned witnesses.
To verify the signature we need to know what is the chunk producer
who should produce a witness at some height, and for that we need
to know what's the epoch_id at this height.