fix log message when a witness is ejected
In previous commit I added debug spans to avoid
writing witness_height, ... everywhere. I also
removed them from the ejection message, but that
was a mistake. The span contains information about
the witness that is being added, not the one that
is being ejected. Because of that the message about
ejecting a witness would contain invalid witness_height
and other parameters. Fix by using the correct parameters,
and names which don't collide with the ones in the span.