fix(mirror): change the default extra key when running with a key mapping secret (#10619)
When records are prepared for a forked network to work with the mirror
tool, there is an option to include a secret that will be used to map
keys so that only someone with the secret can generate the keys in the
forked chain. There is also a default key that is added to any account
that doesn't have a full access key in the source chain so that we can
send transactions on its behalf in the forked chain. But there was an
oversight where this default extra key is not mixed with the secret,
which means that anyone can control any account that doesn't have any
full access key in the source chain. So fix it by mapping the key with
the secret, and leaving it as-is when there's no secret, which should be
backward compatible with old setups with no secret.

This hasn't been an issue, but the records prepared for the statelessnet
program were prepared with a secret so that only the test operator can
control the accounts coming from mainnet, and these accounts were
therefore exposed on statelessnet. They have since all been updated
manually to use the default key defined in this PR