fix: Don't ban peers for invalid ChunkStateWitness messages (#10667)
The current logic of banning peers that sent a message with invalid
ChunkStateWitness is flawed - it bans the peer that forwarded the
message instead of the one that produced it, which means that innocent
peers could be banned for forwarding a bad message.

Let's remove the existing banning logic because it's invalid. Later it
can be properly reimplemented in another PR.

Fixes: https://github.com/near/stakewars-iv/issues/17
Refs: [zulip
conversation](https://near.zulipchat.com/#narrow/stream/407237-pagoda.2Fcore.2Fstateless-validation/topic/StatelessNet.20monitoring/near/423565210)