[TestLoop Refactoring] (4/n) Further refactorings to allow ClientActions to be run in a TestLoop test (#10661)
A few changes:
- Split SyncJobsActor into SyncJobsActions/SyncJobsActor, the same way
it was done for the ClientActor.
- Replace the `Box<dyn Fn(...)>` with `Sender<...>` for the
SyncJobsActor message senders given to ClientActions and other functions
- Replace the sync jobs `ArbiterHandle` with a `FutureSpawner`. Added a
`ActixArbiterHandleFutureSpawner` that spawns the future onto the
`ArbiterHandle` rather than the spawner's arbiter.
 - Replace `Addr<TelemetryActor>` with `Sender<TelemetryEvent>`.
- Introduced some test utils for making event listeners for
TestLoop-based tests that use ClientActions and SyncJobsActions.
- A few minor improvements for the TestLoop API, and the log visualizer
UI.
- A WIP example on testing Client with TestLoop. It doesn't do anything
useful yet, but it's an attempt to get started. The major TODO here is
to make the network layer work.