fix(resharding): create flat storage for children shards after node restart (#10684)
The root cause of the resharding issue in mocknet testing seems to be
that when the node is restarted the flat storage is not recreated.
Without flat storage when applying the first block of the new epoch we
get different state roots than the nodes that do have flat storage.

This PR introduces creating the flat storage for the children shards aka
next epoch shards. This is depended on the flat storage status that is
set at the end of resharding
[here](https://github.com/near/nearcore-private/blob/1a4756e1acbebc073f312fde1457be47e1da3fc0/chain/chain/src/resharding.rs#L180).

I failed to reproduce the issue so we'll need to test it in mocknet
unfortunately. We should test all the cases (no restart, restart during
resharding, restart during catchup, restart post catchup). cc
@marcelo-gonzalez

Ideally we should also error out when trying to apply a chunk without
flat storage but I'm not brave enough to do it in 1.37 :)

I also changed shard_id to shard_uid in a few places but this needs to
happen on a wider scale. I'll do that separately too.