Standardize most of the codebase to use near_async::time. (#10677)
Replace the uses of
 * `chrono::{Instant, Duration, DateTime<Utc>}`
 * `std::time::{Duration, Instant}`
 * `StaticClock`
with `near_async::time::{Duration, Instant, Utc, Clock}`, the first
three of which are aliases for `time::{Duration, Instant,
OffsetDateTime}`.

Why?
* It's very inflexible to use several time libraries and then having to
convert between them.
* `StaticClock` is terrible because it's static, it's not even possible
to directly adapt TestLoop to also affect `StaticClock`, because
`StaticClock` is thread-local, lol. But so many places use
`StaticClock`, making fake-time-based integration testing impossible, so
it's necessary to get rid of it.
 * `StaticClock` uses chrono, but `Clock` uses `near_async::time::...`

The replacement involves the following:
* Introduce the use of Clock for Client, Chain, Doomslug,
BlockDelayTracker, etc., and remove parameters that used to take
timestamps directly and use the clock instead, replace usages of
StaticClock with this clock, and replace usages of .elapsed() with a
subtraction from clock.now().
* This replacement is almost complete but there are still remaining
pieces that use .elapsed(), ::now(), etc., but they are not critical and
will be addressed later.
 * `StaticClock` is completely deleted.
* The `near_async::time::Duration` is signed, whereas
`std::time::Duration` is unsigned. So some care is taken to preserve
existing behavior. `a.saturating_duration_since(b)` is replaced with `(a
- b).max(Duration::ZERO)` or depending on context, an equivalent form
such as `(a - b).whole_milliseconds().max(0) as u64` or `(a -
b).as_seconds_f64().max(0.0)`.
* Some APIs requires `std::time::Duration`, such as tokio::time::sleep,
and aws APIs. This requires a conversion. When it is obvious that the
source signed duration is non-negative, I directly use `unsigned_abs()`;
otherwise, I use a `.max(Duration::ZERO).unsigned_abs()`.
* Some serialized structs must continue to use std::time::Duration,
notably `GenesisConfig` and `Config`. For the latter, I've still
switched to using `near_async::time::Duration`, but added a serde-with
annotation so that they are serialized as `std::time::Duration`. An
additional test is added to assert that all `Duration`s are annotated
this way. This is important in ensuring that the config.json files
remain compatible.
* from_timestamp and to_timestamp is replaced with
`Utc::from_unix_timestamp_nanos` and `unix_timestamp_nanos`.

There's also a side change to add more to the
simple_test_loop_example.rs test.