refactor: erase ReceiptEnum::is_action and ::action (#10698)
In #10676 we introduced ReceiptEnum variants `PromiseYield` and
`PromiseResume`.

At the time `Action` and `PromiseYield` receipts were treated
identically through the protocol, so it was reasonable to have unified
helpers `is_action` and `action`.

After #10689 there are fundamental and subtle differences between
`Action` receipts and`PromiseYield` receipts (for example how they are
handled in
[apply_receipt](https://github.com/near/nearcore/blob/6a68a2b18a8e17c9f1946d5ccd49742e7e14b587/runtime/runtime/src/lib.rs#L1039)).
It seems safer now to explicitly specify behavior everywhere.