yield/resume: rework receipt manager (#10702)
By commit:
1. Renames receipt creation functions of the ReceiptManager and in
downstream files for clarity.
2. Refactors action receipt representation in ReceiptManager from
`(AccountId, ReceiptMetadata)` by moving the receiver account id inside
the metadata struct.
3. Similarly refactors data receipt representation from tuple
`(CryptoHash, Vec<u8>)` into a struct.
4. Changes behavior in the case that a promise yield is resumed within
the same transaction in which it is created. This case is now handled by
converting the `PromiseYield` to a standard `Action` receipt and
delivering the resume payload as a standard `Data` receipt.

(4) is important because unlike `PromiseYield` and `PromiseResume`
receipts, standard `Action` and `Data` receipts are robust to delivery
order (see
[process_receipt](https://github.com/near/nearcore/blob/00d51bde97378995d2adc058fa30fe299e18dc38/runtime/runtime/src/lib.rs#L898-L1094)).
By making this change, we lift an implicit requirement that the local
receipts created as a result of a single transaction would be applied in
the order in which they were created.