keep the lock for all of download_binaries()
unclear why this was even written to take the lock at different
points inside this function anyway. originally it was only called
at startup at which point there wouldn't be any contention anyway.

this means that the update_binaries rpc will block all other rpc
methods, so that should be fixed. But it was previously sort of unsound
if we get concurrent calls to update_binaries