fix(mocknet): avoid calling sys.exit() in neard runner
sys.exit() doesn't do what you'd think in python threading.Threads.
If it's called in an RPC handler, it just causes that handler to
exit, and when it's called in main_loop(), it just causes that function
to exit, but the program is still running and accepting connections, which
can be quite confusing when main_loop() was expected to do some work after
an RPC call returns OK

so instead of calling sys.exit(), we can add a TestState.ERROR variant
and set the state to that when something bad seems to have happened,
and then the test operator can take a look and hopefully fix it.