[stateless_validation] Add a soft size limit for state witness (#10703)
This PR adds a new runtime config `state_witness_size_soft_limit` with
size about 16 MB to begin with along with an implementation to enforce
it in runtime.

This is the first step of https://github.com/near/nearcore/issues/10259

What is state witness size soft limit?

In order to limit the size of the state witness, as a first step, we are
adding a limit to the max size of the state witness partial trie or
proof. In runtime, we record all the trie nodes touched by the chunk
execution and include these in the state witness. With the limit in
place, if the size of the state witness exceeds 16 MB, then we would
stop applying receipts further and push all receipts into the delayed
queue.

The reason we call this a soft limit is that we stop the execution of
the receipts AFTER the size of the state witness has exceeded 16 MB.

We are including this as part of a new protocol version 83

Future steps
- Introduce limits on other parts of the state witness like new
transactions
- Introduce a hard size limit for individual contract executions
- Monitor size of state witness
- Add metrics in a separate PR