introduce an always-failing fuzzer, to sanity-check our clusterfuzz setup (#10723)
Based on the clusterfuzz web UI I cannot confirm whether it is actually
working properly, and I’m under the impression that it is actually not
running the fuzzers we’re asking it to.

This fuzzer should make it obvious: if clusterfuzz does not report a
fuzzer failure on it, then it is broken.

There is no good place for it, so I just put it in our current
"everything else" near-primitives crate, as it has basically no internal
dependencies on anything it’ll be very easy to move it out whenever we
actually get close to getting rid of near-primitives.