fix(epoch-manager): save all validator production stats (#10683)
The HashMap returned by `EpochManager::compute_kickout_info()` currently
removes all stats for validators who were kicked for not producing
enough blocks before returning. This is required because
`RewardCalculator::calculate_reward()` requires that the
`validator_block_chunk_stats` arg passed to it only contain non-kicked
validators that get rewards. But an unintended consequence of this is
that we don't save to disk the stats for kicked validators, and calls to
the "validators" RPC method will return all zeros for those validators.

Since the only two places where this validator stats map is accessed is
in responding to those RPC calls and computing rewards, we can change it
to not to forget about kicked validators as long as we remove those
before computing rewards.

To make sure this is correct, we need to make sure that this removal of
validator stats by checking whether the kickout reason is one of
`NotEnoughBlocks` or `NotEnoughChunks` exactly matches the removals we
previously got from `compute_kickout_info()`. That function only inserts
kickouts of those two types, and the `validator_kickout` variable it's
merged into in `EpochManager::collect_blocks_info()` only inserts
kickouts of other types.