yield/resume: init integration tests (#10746)
This PR adds a couple of basic integration tests for yield execution:

- tests behavior in the case `promise_yield_resume` is called with an
unexpected data id
- tests behavior in the case that `promise_yield_create` and
`promise_yield_resume` are called within the same transaction

Why do we start with these edge cases? As it stands the testing
infrastructure in
[runtime_user.rs](https://github.com/near/nearcore/blob/master/integration-tests/src/user/runtime_user.rs#L344)
expects all receipts generated by a transaction to be resolved
immediately once the transaction is executed. Calling yield without
resume in any transaction breaks that assumption, so adding such tests
will be a bit more involved.

I had to drop the placeholder gas costs to get these tests to pass.

I will follow up with a PR which tweaks RuntimeUser and adds tests for
the main cases:

- yield in one transaction, followed by resume in a later one
- yield without resume, leading to a timeout