congestion: print queue size in bytes and gas (#10752)
So far, we only printed the queue length, as in "number of elements".
But we are much more interested in the size in bytes, as this is the
real constraint. Also add the same for attached gas, as this can be
interesting, too.

The way this is computed, it's rather inefficient. We might want to
compute the sums more efficiently.
But running

`time cargo run --release -- --write-stats`

still finishes in
`real    0m18.183s`

on my machine, so I guess it's still tolerable, for now...