vm: store compiled contracts directly in the filesystem
Going through rocksdb involves compression, all the rocksdb logic etc.
That can't be cheap. This is also a cache so none of the features really
benefit the use-case.

Storing compiled contracts as files also makes it much more
straightforward to inspect them now if the need to do so arises. For
example an estimation of the size distribution is one `ls` away.