Attempt to thread the FilesystemCompiledContractCache through to configs
Creating the cache requires us to know the right place to create it at.
This has required me to thread the thing pretty much though the entire
stack. But it also has resulted in some benefits -- for instance we
won't be unnecessarily doing any directory related operations every time
a view call is invoked cause the cache is created pretty early on and is
reused.

This is also a step in the right direction in the longer term where we
probably want the contract runtime to stand on its own a little more and
have the invoker of the transaction runtime to be responsible for
integrating the components together.