fix(docs): Fix an "instrument" link in the docs (#10804)
fixing one of the lychee failures

The other one is mysterious, it works for me but not for lychee. I do
not have the strenght to look into it now.

I'm going to mark this PR merge when ready, please remove it from the
queue if you have any feedback.