fix(snapshots): drop unwanted column families instead of deleting all keys (#10803)
If the `columns_to_keep` arg of
`checkpoint_hot_storage_and_cleanup_columns()` is `Some`, then we delete
all the data in every other column. Then if snapshot compaction is
enabled in the configs, we rely on that to clean up the files on disk.
Instead of doing that, we can just call `drop_cf()` on every unwanted
column family, and the associated sst files will be removed without the
need for any compactions.

So this moves the `columns_to_keep` arg to
`near_store::db::Database::create_checkpoint()`, and has the rocksdb
implementation of that trait call `drop_cf()` on unwanted column
families. These column families are then immediately recreated in
`checkpoint_hot_storage_and_cleanup_columns()` by the call to
`StoreOpener::open_in_mode()`, but the data on disk is gone.

This also means we can get rid of the state snapshot options in the
config, since they were only ever intended to clean up the unwanted
files, which aren't there anymore.