congestion: report transaction delay stats (#10821)
The goal is to measure user experience.

1. Measure the time between when a transaction is first submitted until
it finishes, either successfully or by being rejected.
2. Count how many transactions are unresolved, as this means a user does
not know yet what happens to the transaction.

In the summary table, I only print the mean successful transaction
delay.

For the CSV writer, we have mean, average, and 90th percentiles of
successful and rejected transactions reported separately. Plus the
number of unresolved transactions. A combination of these numbers in
graphs should help to understand user experience in reasonable details.