fix: remove direct indexing by shard id (#10831)
Removed some of the usage of [shard_id] that can end up with "out of
bounds" panics. This is both to better handle malicious inputs and shard
layout changes during resharding. Ideally I'd like to remove the Index
trait on ChunkCollection but I'm not quite there yet. Deprecating it
doesn't work because it's a trait...