near_vm: change LimitedMemoryPool to create new maps if dry
Now that the size of the UniversalArtifact cache is configurable by the
operator, the LimitedMemoryPool would need to account for the
configuration and allocate an appropriate number of pages depending on
that configuration. However, we do not have easy access to that
configuration from the contract runtime, and I don't think I want to
unravel the thread of moving the set up of contract runtime resources
all the way down to the place where this configuration is directly
available.

In the long term we should still do it (and consider reverting the pool
to be "limited"), but for the time being this will have to do.