near_vm: change LimitedMemoryPool to create new maps if dry (#10839)
Now that the size of the `UniversalArtifact` cache is configurable by
the operator, the `LimitedMemoryPool` would need to account for the
configuration and allocate an appropriate number of pages depending on
that configuration. However, we do not have easy access to that
configuration from the contract runtime, and I don't think I want to
unravel the thread of moving the set up of contract runtime resources
all the way down to the place where this configuration is directly
available.

In the long term we should still keep memory pool limited somehow and
make sure we pre-allocate exactly the required number of maps of the
right size. To do that we would have to refactor the contract runtime
setup to rely less on the global state and do it somewhere near the
setup of the transaction runtime as well. This is also relevant wrt.
limited replayability work.