Always record storage when stateless validation is enabled (#10859)
For stateless validation we need to record all trie reads performed when
applying a chunk in order to prepare a `PartialState` that will be
included in `ChunkStateWitness`

Initially it was only necessary to record trie reads when producing a
chunk. Validators could read all the values from the provided
`PartialState` without recording anything.

A recent change (https://github.com/near/nearcore/pull/10703) introduced
a soft limit on the size of `PartialState`. When applying a chunk we
watch how much state was recorded, and once the amount of state exceeds
the soft limit we stop applying the receipts in this chunk.
This needs to be done on both the chunk producer and chunk validator -
if the chunk validator doesn't record reads and enforce the limit, it
will produce a different result of chunk application, which would lead
to validation failure.

This means that state should be recorded in all cases when a statelessly
validated chunk is applied. Let's remove the configuration option that
controls whether trie reads should be recorded (`record_storage`) and
just record the reads on every chunk application (when
`statelessnet_protocol` feature is enabled).

Refs: [zulip
conversation](https://near.zulipchat.com/#narrow/stream/407237-core.2Fstateless-validation/topic/State.20witness.20size.20limit/near/428313518)