fix(resharding) - fixed resharding rpc tx test (#10857)
The test waits for transactions to get fully processed which slows it
down. I extended the epoch lenght so that we can comfortably prepare and
send transactions before resharding and check the status after.

There was a small bug in the epoch boundary calculation that wasn't
exposed until now because we would always enter the loop late.