runtime: continue prefetching between receipts
Some prefetcher requests can take a longer time to execute than the
receipt being processed. So clearing the queue between all the receipts
can lead to prefetching of account data that we prefetch for finalize
not happen for the most part, as the prefetcher being busy with its
current task wouldn't even see the prefetch request.

This is a pretty crude implementation: ideally we'd have some sort of
priority queue here, but for now it works alright.