Split storage on genesis (#10906)
#10102

Adding the ability for cold storage node to start from genesis without
gaps in data or crashes in cold loop.

1. Making split storage node a default archival node via changing
`get_default_kind`
2. Adding tool to make DB specified in `cold_store` part of the config
considered empty on startup. This is needed for forknet. Changes to
`core/store/src/db/colddb.rs` are due to this tool.
3. **Hard and important part**. Adding genesis migration to cold loop.
Genesis migration is basically the same initial migration that we use
for legacy archival nodes (copy all cold columns), but we need to set a
different head. So, now we have migration loop that first determines
what cold head to set to cold storage based on DBKind of hot store, and
then performs full cold copy.