fix(mocknet): clear old binaries from data.json in update-binaries
When the `force` arg of download_binaries() is true, we want to
redownload everything, but the way it's done before this PR is by
appending info on the new binaries downloaded to the existing "binaries"
field in data.json. This is not correct because it will mess up the mid-test
binary upgrade logic. So just make sure to clear that field instead of appending
to the existing one

Also in the future we could consider making this a bit easier to use by passing
the URLs to download from in the RPC arguments, which wasn't done originally because
the ports were open to the internet. Now that it's behind ssh this isn't a concern