fix(stateless-validation) - fix protocol version checks (#10933)
The checked_feature macro should always be used with the protocol
version of the current block, not the binary protocol version. Otherwise
the new feature would take effect during binary upgrade instead of
protocol upgrade. If that happens all sorts of trouble may happen as
nodes using different binaries may get diverge e.g. in state roots.

This is a preparation for safe mainnet release.

I know it was under discussion previously but even if in this particular
case it doesn't matter I would feel much safer having it done the proper
way. Also it's all to easy to copy paste it elsewhere when working on
the next feature so best to just stick to the rules.

I picked the StateWitnessSizeLimit as it seemed like the proper
ProtocolFeature to use for this, please correct me if I'm wrong. It
doesn't really matter now though.