rpc: partial revert of #10792 (#10951)
This PR is based on https://github.com/near/nearcore/pull/10948 because
it touches the same places
It's better to look at the separate commits

So first commit can be reviewed separately in
https://github.com/near/nearcore/pull/10948

Second commit is a partial revert of
https://github.com/near/nearcore/pull/10792
I just returned this piece back, no changes is made

https://github.com/near/nearcore/pull/10792/files#diff-ef9c6aaa80a330e446c5365f42be9bff37ba4f898cf519dadd7e17545783c77cL3099-L3126

Third commit is the meaningful part. We need to adjust the logic with
the old implementation.
1. We now treat incomplete list of the receipts as a valid state. It
means no `expect` usage, handle empty list of outcomes
2. We no longer can use `status` field for `finalExecutionStatus`