Feat: Eth wallet contract implementation (#10850)
This PR adds the eth wallet contract implementation which will be
automatically deployed with eth implicit accounts. This is as per the
design outlined in https://github.com/near/NEPs/issues/518 .

It is intentional that the wallet contract implementation is isolated as
its own crate and not contained in the broader nearcore workspace. This
allows it to be reviewed, tested and audited independently from the rest
of the Near code. The crate includes integration tests written using
`near-workspaces`. However, the nearcore integration test related to the
wallet contract has also been updated to check that the contract is
automatically deployed when creating an eth address and that it works as
expected.

This will not be the last PR in this project because I am only adding
the logic for the implementation. It still remains to setup the
reproducible build pipeline and update the tests which check the
contract hashes. Additionally, there are still a few details which need
to be finalized: (1) the Ethereum chain ID that will be associated with
Near and (2) the Near account ID that will have the eth address
registrar contract deployed. The eth address registrar contract stores a
reverse lookup of Ethereum-like address to Near account IDs. It is
necessary for the wallet contract to detect faulty relayers.

However this (relatively large) PR can be reviewed and merged, and the
points above will be addressed in much more manageable follow-up PRs.