Fix update cache (#10973)
There is a lock contention issue on `caches` in `ShardTries`. We always
acquire the write lock for the `Rwlock` but that is only needed when
there is a change in the number of shards, which happens extremely
rarely. This lock contention partially leads to #10970. The fix here
works by always checking whether the shard already exists and only
acquire a write lock if the shard does not exist and the outer hash map
needs to be modified.