congestion: transaction pool trimming (#11024)
If we run with infinite transaction pools, some workloads report really
long transaction delays. This reflects reality in some sense. But is not
really useful for comparison.

The CLI flag `tx_pool_size` allows to trim the pool, possibly even empty
it completely. Then, the reported delays are shorter because started
transactions are fresher.