o11y: better default values for otel batch processor
With the default-default values we're having a hard time keeping up with
publishing traces even on some of the most strict filters due to the
fact that our spans are pretty bursty.

This leads to a large number of

    OpenTelemetry trace error occurred.
    cannot send span to the batch span processor because the channel is
    full

being output to the standard output. This attempts to mitigate it.