Expose information about the size of recorded trie storage to the runtime
To enforce the hard per-receipt limit we need to monitor how much storage
proof has been recorded during execution of the receipt and halt the
execution when the size of generated storage proof goes over the limit.

To achieve this the runtime needs to be able to see how much proof was recorded,
so let's expose this information so that it's available from the runtime.

`recorded_storage_size()` doesn't provide the exact size of storage
proof, as it doesn't cover some corner cases (see https://github.com/near/nearcore/issues/10890),
so we use the `upper_bound` version to estimate how much storage proof
could've been generated by the receipt. As long as upper bound is
under the limit we can be sure that the actual value is also
under the limit.