Track size of the recorded storage proof size during contract execution
Add a `RecordedStorageCounter` which functions similarly to `GasCounter`.
After every trie operation (`storage_write`, `storage_read`, ...) the
runtime calls `RecordedStorageCounter::observe_size` to keep track
of the size of recorded storage proof. If the size exceeds the allowed
limit, then the receipt execution failed, just like it would fail
if the receipt went over the allowed gas limit.

For now the limit is `usize::MAX`, the proper limit will be set
in a following commit.